pyproject.toml
[tool.poetry]
name = "wordgame"
version = "0.1.0"
description = ""
authors = ["rauhja <82282638+rauhja@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
invoke = "^1.7.3"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
coverage = "^6.5.0"
pylint = "^2.15.6"
autopep8 = "^2.0.0"
python-dotenv = "^0.21.0"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
